pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "newsapi-python>=0.2.7",
    "pandas>=2.2.3",
    "plotly>=6.0.0",
    "requests>=2.32.3",
    "streamlit>=1.42.2",
    "trafilatura>=2.0.0",
    "yfinance>=0.2.54",
]
